Add new file structure for Retrofit
