+++ app/src/main/java/com/blockeq/stellarwallet/activities/WalletActivity.kt
@@ -13,7 +13,7 @@ import com.blockeq.stellarwallet.fragments.WalletFragment
 import com.blockeq.stellarwallet.interfaces.OnLoadAccount
 import com.blockeq.stellarwallet.interfaces.OnLoadEffects
 import com.blockeq.stellarwallet.models.MinimumBalance
-import com.blockeq.stellarwallet.remote.Horizon
+import com.blockeq.stellarwallet.networking.Horizon
 import com.blockeq.stellarwallet.utils.AccountUtils
 import com.blockeq.stellarwallet.utils.KeyboardUtils
 import com.blockeq.stellarwallet.utils.NetworkUtils
